+++ js/main.js
@@ -1,2 +1,5 @@
-import './functions.js';
-import './data-generate.js';
+import {similarPictures} from './data-generate.js';
+import {stringCheck, getPalindrome} from './util.js';
+
+console.log(similarPictures)
+
