+++ js/util.js
@@ -4,5 +4,21 @@ const getRandom = (a, b) => {
   const result = Math.random() * (upper - lower + 1) + lower;
   return Math.floor(result);
 };
+const stringCheck = (string, maxLenth) => {
+  const check = string.length <= maxLenth;
+  console.log(check);
+};
 
-export {getRandom};
+const getPalindrome = (input) => {
+  const normalizedStr = input.replaceAll(' ', '').toLowerCase();
+
+  let reversedStroke = '';
+
+  for (let i = normalizedStr.length - 1; i >= 0; --i) {
+    reversedStroke += normalizedStr[i];
+  }
+
+  return normalizedStr === reversedStroke;
+};
+
+export {getRandom, stringCheck, getPalindrome};
